backend/requirements.txt
flask
flask-cors
google-generativeai
psycopg2-binary
python-dotenv
